Controller Evcs: cleanup Split impl and interface
