+++ io.openems.edge.controller.evcs/test/io/openems/edge/controller/evcs/EvcsControllerImplTest.java
@@ -2,6 +2,9 @@ package io.openems.edge.controller.evcs;
 
 import static org.junit.Assert.assertEquals;
 
+import java.time.Instant;
+import java.time.ZoneOffset;
+
 import org.junit.Test;
 
 import io.openems.common.types.ChannelAddress;
@@ -10,24 +13,25 @@ import io.openems.edge.common.sum.DummySum;
 import io.openems.edge.common.test.AbstractComponentTest.TestCase;
 import io.openems.edge.common.test.DummyComponentManager;
 import io.openems.edge.common.test.DummyConfigurationAdmin;
+import io.openems.edge.common.test.TimeLeapClock;
 import io.openems.edge.controller.test.ControllerTest;
 import io.openems.edge.evcs.api.ChargeMode;
 import io.openems.edge.evcs.api.Status;
 import io.openems.edge.evcs.test.DummyEvcsPower;
 import io.openems.edge.evcs.test.DummyManagedEvcs;
 
-public class EvcsControllerTest {
+public class EvcsControllerImplTest {
 
 	private static final DummyEvcsPower EVCS_POWER = new DummyEvcsPower(new DisabledRampFilter());
 	private static final DummyManagedEvcs EVCS = new DummyManagedEvcs("evcs0", EVCS_POWER);
 
-	private static String EVCS_ID = EVCS.id();
-	private static boolean ENABLE_CHARGING;
-	private static ChargeMode CHARGE_MODE;
-	private static int FORCE_CHARGE_MIN_POWER = 7360;
-	private static int DEFAULT_CHARGE_MIN_POWER = 0;
-	private static Priority PRIORITY = Priority.CAR;
-	private static int ENERGY_SESSION_LIMIT = 0;
+	private static final String EVCS_ID = EVCS.id();
+	private static final boolean DEFAULT_ENABLE_CHARGING = true;
+	private static final ChargeMode DEFAULT_CHARGE_MODE = ChargeMode.EXCESS_POWER;
+	private static final int DEFAULT_FORCE_CHARGE_MIN_POWER = 7360;
+	private static final int DEFAULT_CHARGE_MIN_POWER = 0;
+	private static final Priority DEFAULT_PRIORITY = Priority.CAR;
+	private static final int DEFAULT_ENERGY_SESSION_LIMIT = 0;
 
 	private static ChannelAddress sumGridActivePower = new ChannelAddress("_sum", "GridActivePower");
 	private static ChannelAddress sumEssDischargePower = new ChannelAddress("_sum", "EssDischargePower");
@@ -43,11 +47,7 @@ public class EvcsControllerTest {
 	@Test
 	public void excessChargeTest1() throws Exception {
 
-		ENABLE_CHARGING = true;
-		CHARGE_MODE = ChargeMode.EXCESS_POWER;
-		PRIORITY = Priority.CAR;
-
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -55,12 +55,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(DEFAULT_ENABLE_CHARGING) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(DEFAULT_FORCE_CHARGE_MIN_POWER) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.input(sumEssDischargePower, -6000) //
@@ -74,10 +74,7 @@ public class EvcsControllerTest {
 	@Test
 	public void excessChargeTest2() throws Exception {
 
-		ENABLE_CHARGING = true;
-		CHARGE_MODE = ChargeMode.EXCESS_POWER;
-		PRIORITY = Priority.STORAGE;
-		final var test = new ControllerTest(new EvcsController()) //
+		final var test = new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -85,12 +82,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(DEFAULT_ENABLE_CHARGING) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(DEFAULT_FORCE_CHARGE_MIN_POWER) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(Priority.STORAGE) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()); //
 
 		test.next(new TestCase() //
@@ -106,12 +103,7 @@ public class EvcsControllerTest {
 	@Test
 	public void forceChargeTest() throws Exception {
 
-		ENABLE_CHARGING = true;
-		FORCE_CHARGE_MIN_POWER = 7360;
-		CHARGE_MODE = ChargeMode.FORCE_CHARGE;
-		PRIORITY = Priority.CAR;
-
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -119,12 +111,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(DEFAULT_ENABLE_CHARGING) //
+						.setChargeMode(ChargeMode.FORCE_CHARGE) //
+						.setForceChargeMinPower(DEFAULT_FORCE_CHARGE_MIN_POWER) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.input(sumEssDischargePower, -5000) //
@@ -137,9 +129,7 @@ public class EvcsControllerTest {
 	@Test
 	public void chargingDisabledTest() throws Exception {
 
-		ENABLE_CHARGING = false;
-
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -147,12 +137,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(false) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(DEFAULT_FORCE_CHARGE_MIN_POWER) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.output(evcs0SetChargePowerLimit, 0));
@@ -161,11 +151,8 @@ public class EvcsControllerTest {
 	@Test
 	public void wrongConfigParametersTest() throws Exception {
 
-		DEFAULT_CHARGE_MIN_POWER = 30000;
-		FORCE_CHARGE_MIN_POWER = 30000;
-
 		var cm = new DummyConfigurationAdmin();
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", cm) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -173,12 +160,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
-						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setEnableCharging(DEFAULT_ENABLE_CHARGING) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(30_000) //
+						.setDefaultChargeMinPower(30_000) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.input(evcs0MaximumHardwarePower, 12000));
@@ -190,25 +177,24 @@ public class EvcsControllerTest {
 	@Test
 	public void clusterTest() throws Exception {
 
-		ENABLE_CHARGING = true;
-		FORCE_CHARGE_MIN_POWER = 3333;
-		CHARGE_MODE = ChargeMode.EXCESS_POWER;
-		PRIORITY = Priority.CAR;
+		final var clock = new TimeLeapClock(Instant.ofEpochSecond(1577836800) /* starts at 1. January 2020 00:00:00 */,
+				ZoneOffset.UTC);
+		final var componentManager = new DummyComponentManager(clock);
 
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
-				.addReference("componentManager", new DummyComponentManager()) //
+				.addReference("componentManager", componentManager) //
 				.addReference("sum", new DummySum()) //
 				.addReference("evcs", EVCS) //
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(DEFAULT_ENABLE_CHARGING) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(3_333) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.input(sumEssDischargePower, -10000) //
@@ -245,12 +231,7 @@ public class EvcsControllerTest {
 	@Test
 	public void clusterTestDisabledCharging() throws Exception {
 
-		ENABLE_CHARGING = false;
-		FORCE_CHARGE_MIN_POWER = 3333;
-		CHARGE_MODE = ChargeMode.EXCESS_POWER;
-		PRIORITY = Priority.CAR;
-
-		new ControllerTest(new EvcsController()) //
+		new ControllerTest(new EvcsControllerImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("componentManager", new DummyComponentManager()) //
 				.addReference("sum", new DummySum()) //
@@ -258,12 +239,12 @@ public class EvcsControllerTest {
 				.activate(MyConfig.create() //
 						.setId("ctrlEvcs0") //
 						.setEvcsId(EVCS_ID) //
-						.setEnableCharging(ENABLE_CHARGING) //
-						.setChargeMode(CHARGE_MODE) //
-						.setForceChargeMinPower(FORCE_CHARGE_MIN_POWER) //
+						.setEnableCharging(false) //
+						.setChargeMode(DEFAULT_CHARGE_MODE) //
+						.setForceChargeMinPower(3_333) //
 						.setDefaultChargeMinPower(DEFAULT_CHARGE_MIN_POWER) //
-						.setPriority(PRIORITY) //
-						.setEnergySessionLimit(ENERGY_SESSION_LIMIT) //
+						.setPriority(DEFAULT_PRIORITY) //
+						.setEnergySessionLimit(DEFAULT_ENERGY_SESSION_LIMIT) //
 						.build()) //
 				.next(new TestCase() //
 						.input(sumEssDischargePower, -10000) //
